requirements.txt
paho-mqtt
RPLCD
smbus2
config
